# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Reality Compiler API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /marketplace/listings:
    get:
      operationId: listMarketplaceListings
      tags: [marketplace]
      summary: List published marketplace designs
      parameters:
        - in: query
          name: category
          required: false
          schema:
            type: string
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [popular, price-asc, price-desc, newest]
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceListingSummary"
    post:
      operationId: publishListing
      tags: [marketplace]
      summary: Publish a session as a marketplace listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishListingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceListingDetail"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not session owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /marketplace/listings/{id}:
    get:
      operationId: getMarketplaceListing
      tags: [marketplace]
      summary: Get a marketplace listing detail with embedded design output
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceListingDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: unpublishListing
      tags: [marketplace]
      summary: Unpublish a marketplace listing
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /marketplace/profile/{userId}:
    get:
      operationId: getDesignerProfile
      tags: [marketplace]
      summary: Get a designer profile (listings + total orders)
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DesignerProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    BomItem:
      type: object
      properties:
        component:
          type: string
        material:
          type: string
        quantity:
          type: number
        unit:
          type: string
        unitCost:
          type: number
        totalCost:
          type: number
      required:
        - component
        - material
        - quantity
        - unit
        - unitCost
        - totalCost
    CostEstimate:
      type: object
      properties:
        low:
          type: number
        high:
          type: number
        currency:
          type: string
        leadTimeDays:
          type: integer
      required:
        - low
        - high
        - currency
        - leadTimeDays
    DesignOutput:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        productName:
          type: string
        category:
          type: string
        summary:
          type: string
        primaryMaterial:
          type: string
        materials:
          type: array
          items:
            type: string
        dimensions:
          type: string
        weightGrams:
          type: ["number", "null"]
        processes:
          type: array
          items:
            type: string
        bom:
          type: array
          items:
            $ref: "#/components/schemas/BomItem"
        costEstimate:
          $ref: "#/components/schemas/CostEstimate"
        imageUrl:
          description: URL to a hosted image (object storage URL or null)
          type: ["string", "null"]
        manufacturingNotes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - productName
        - category
        - summary
        - primaryMaterial
        - materials
        - dimensions
        - processes
        - bom
        - costEstimate
        - manufacturingNotes
        - createdAt
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        location:
          type: string
        country:
          type: string
        capabilities:
          type: array
          items:
            type: string
        materials:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        leadTimeMinDays:
          type: integer
        leadTimeMaxDays:
          type: integer
        pricingMultiplier:
          type: number
        setupFee:
          type: number
        rating:
          type: number
        capacityLevel:
          type: string
          enum: [low, medium, high]
      required:
        - id
        - slug
        - name
        - tagline
        - description
        - location
        - country
        - capabilities
        - materials
        - certifications
        - leadTimeMinDays
        - leadTimeMaxDays
        - pricingMultiplier
        - setupFee
        - rating
        - capacityLevel
    ProcessBreakdownItem:
      type: object
      properties:
        process:
          type: string
        description:
          type: string
        cost:
          type: number
      required:
        - process
        - description
        - cost
    QuoteScoreFactors:
      type: object
      properties:
        processMatch:
          type: number
        materialMatch:
          type: number
        leadTime:
          type: number
        rating:
          type: number
        total:
          type: number
      required:
        - processMatch
        - materialMatch
        - leadTime
        - rating
        - total
    Quote:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        supplier:
          $ref: "#/components/schemas/Supplier"
        unitCost:
          type: number
        setupFee:
          type: number
        totalCost:
          type: number
        leadTimeDays:
          type: integer
        processBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/ProcessBreakdownItem"
        scoreFactors:
          $ref: "#/components/schemas/QuoteScoreFactors"
        rank:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - supplier
        - unitCost
        - setupFee
        - totalCost
        - leadTimeDays
        - processBreakdown
        - scoreFactors
        - rank
        - notes
        - createdAt
...
    PublishListingInput:
      type: object
      properties:
        sessionId:
          type: integer
        title:
          type: string
        category:
          type: string
        description:
          type: string
        listingPrice:
          type: number
      required:
        - sessionId
        - title
        - category
        - description
        - listingPrice
    MarketplaceListingSummary:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        userId:
          type: string
        creatorHandle:
          type: string
        creatorDisplayName:
          type: ["string", "null"]
        creatorAvatarUrl:
          type: ["string", "null"]
        title:
          type: string
        category:
          type: string
        description:
          type: string
        listingPrice:
          type: number
        thumbnailUrl:
          type: ["string", "null"]
        primaryMaterial:
          type: ["string", "null"]
        productName:
          type: ["string", "null"]
        orderCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - userId
        - creatorHandle
        - title
        - category
        - description
        - listingPrice
        - orderCount
        - createdAt
    MarketplaceListingDetail:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        userId:
          type: string
        creatorHandle:
          type: string
        title:
          type: string
        category:
          type: string
        description:
          type: string
        listingPrice:
          type: number
        orderCount:
          type: integer
        designOutput:
          $ref: "#/components/schemas/DesignOutput"
        quotes:
          type: array
          items:
            $ref: "#/components/schemas/Quote"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - userId
        - creatorHandle
        - title
        - category
        - description
        - listingPrice
        - orderCount
        - designOutput
        - quotes
        - createdAt
        - updatedAt
    DesignerProfile:
      type: object
      properties:
        userId:
          type: string
        handle:
          type: string
        displayName:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        listings:
          type: array
          items:
            $ref: "#/components/schemas/MarketplaceListingSummary"
        totalListings:
          type: integer
        totalOrders:
          type: integer
        totalPayouts:
          type: number
      required:
        - userId
        - handle
        - listings
        - totalListings
        - totalOrders
        - totalPayouts
...